requirements.txt
Flask
Flask-Cors
openai==0.28
python-dotenv